+++ src/components/VisualSearchDebugAction/index.ts
@@ -1,0 +1,1 @@
+export * from './VisualSearchDebugAction';
