--- src/components/VisualSearchDebugView/index.ts
@@ -1,1 +1,0 @@
-export * from './VisualSearchDebugView';
