+++ src/pages/images/ItemOverview/ItemGrid/ImageItem/ImageItemActions.tsx
@@ -1,8 +1,6 @@
-import { useState } from 'react';
-import { Bug, ImageIcon, Images, MoreVertical, NotebookPen } from 'lucide-react';
-import { VisualSearchDebugDialog } from '@/components/VisualSearchDebugView';
+import { ImageIcon, Images, MoreVertical, NotebookPen } from 'lucide-react';
+import { VisualSearchDebugAction } from '@/components/VisualSearchDebugAction';
 import { LoadedImage } from '@/model';
-import { useVisualSearchAvailable } from '@/utils/useVisualSearch';
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -26,53 +24,31 @@ interface ImageItemActionProps {
 
 export const ImageItemActions = (props: ImageItemActionProps) => {
 
-  const hasVisualSearch = useVisualSearchAvailable();
-
-  const [showVisualSearchDebug, setShowVisualSearchDebug] = useState(false);
-
-  const onOpenVSDebug = (e: Event) => {
-    e.preventDefault();
-    setShowVisualSearchDebug(true);
-  }
-
   return (
-    <>
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <button
-            className="item-actions-trigger absolute bottom-2 right-1">
-            <MoreVertical size={18} />
-          </button>
-        </DropdownMenuTrigger>
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <button
+          className="item-actions-trigger absolute bottom-2 right-1">
+          <MoreVertical size={18} />
+        </button>
+      </DropdownMenuTrigger>
 
-        <DropdownMenuContent align="start">
-          <DropdownMenuItem onSelect={props.onSelect}>
-            <NotebookPen className="size-4 text-muted-foreground mr-2" /> Metadata
-          </DropdownMenuItem>
+      <DropdownMenuContent align="start">
+        <DropdownMenuItem onSelect={props.onSelect}>
+          <NotebookPen className="size-4 text-muted-foreground mr-2" /> Metadata
+        </DropdownMenuItem>
 
-          <DropdownMenuItem onSelect={props.onOpen}>
-            <ImageIcon className="size-4 text-muted-foreground mr-2" /> Open image
-          </DropdownMenuItem>
+        <DropdownMenuItem onSelect={props.onOpen}>
+          <ImageIcon className="size-4 text-muted-foreground mr-2" /> Open image
+        </DropdownMenuItem>
 
-          <DropdownMenuItem onSelect={props.onAddToWorkspace}>
-            <Images className="size-4 text-muted-foreground mr-2" /> Add to workspace
-          </DropdownMenuItem>
+        <DropdownMenuItem onSelect={props.onAddToWorkspace}>
+          <Images className="size-4 text-muted-foreground mr-2" /> Add to workspace
+        </DropdownMenuItem>
 
-          {hasVisualSearch && (
-            <DropdownMenuItem onSelect={onOpenVSDebug}>
-              <Bug className="size-4 text-muted-foreground mr-2" /> Inspect visual search index
-            </DropdownMenuItem>
-          )}
-        </DropdownMenuContent>
-      </DropdownMenu>
-
-      {hasVisualSearch && (
-        <VisualSearchDebugDialog 
-          image={props.image}
-          open={showVisualSearchDebug}
-          onOpenChange={setShowVisualSearchDebug} />
-      )}
-    </>
+        <VisualSearchDebugAction image={props.image} />
+      </DropdownMenuContent>
+    </DropdownMenu>
   )
 
 }
